+++ app/graphsensedao.py
@@ -1,7 +1,8 @@
 import cassandra.cluster
 from cassandra.query import named_tuple_factory, dict_factory
-import graphsensemodel as gm
 from flask import abort
+import graphsensemodel as gm
+
 
 session = None
 tx_query = {}
@@ -16,6 +17,8 @@ address_transactions_query = {}
 address_transactions_without_limit_query = {}
 address_tags_query = {}
 address_search_query = {}
+label_search_query = None
+label_query = None
 transaction_search_query = {}
 address_cluster_query = {}
 cluster_tags_query = {}
@@ -32,7 +35,7 @@ cluster_addresses_query = {}
 cluster_addresses_without_limit_query = {}
 block_height_query = {}
 statistics_query = {}
-currency_mapping = {}
+keyspace_mapping = {}
 all_exchange_rates = {}
 last_height = {}
 
@@ -50,7 +53,7 @@ def query_exchange_rates(currency, offset, limit):
 
 
 def query_block(currency, height):
-    set_keyspace(session, currency, raw=True)
+    set_keyspace(session, currency, space="raw")
     if height > last_height[currency]:
         abort(404, "Block not available yet")
     result = session.execute(block_query[currency], [height])
@@ -64,7 +67,7 @@ def query_statistics(currency):
 
 
 def query_block_transactions(currency, height):
-    set_keyspace(session, currency, raw=True)
+    set_keyspace(session, currency, space="raw")
     if height > last_height[currency]:
         abort(404, "Block not available yet")
     result = session.execute(block_transactions_query[currency], [height])
@@ -72,11 +75,9 @@ def query_block_transactions(currency, height):
 
 
 def query_blocks(currency, page_state):
-    set_keyspace(session, currency, raw=True)
-    if page_state is not None:
-        page_state = bytes.fromhex(page_state)
-        results = session.execute(blocks_query[currency],
-                                  paging_state=page_state)
+    set_keyspace(session, currency, space="raw")
+    if page_state:
+        results = session.execute(blocks_query[currency], paging_state=page_state)
     else:
         results = session.execute(blocks_query[currency], [10])
     page_state = results.paging_state
@@ -85,7 +86,7 @@ def query_blocks(currency, page_state):
 
 
 def query_transaction(currency, txHash):
-    set_keyspace(session, currency, raw=True)
+    set_keyspace(session, currency, space="raw")
     try:
         rows = session.execute(tx_query[currency], [txHash[0:5], bytearray.fromhex(txHash)])
     except Exception:
@@ -94,12 +95,12 @@ def query_transaction(currency, txHash):
 
 
 def query_transactions(currency, page_state):
-    set_keyspace(session, currency, raw=True)
-    if page_state is not None:
-        page_state = bytes.fromhex(page_state)
+    set_keyspace(session, currency, space="raw")
+    if page_state:
         results = session.execute(txs_query[currency], paging_state=page_state)
     else:
         results = session.execute(txs_query[currency], [10])
+
     page_state = results.paging_state
     transactions = [gm.Transaction(row, query_exchange_rate_for_height(currency, row.height)).__dict__
                     for row in results]
@@ -107,7 +108,7 @@ def query_transactions(currency, page_state):
 
 
 def query_transaction_search(currency, expression):
-    set_keyspace(session, currency, raw=True)
+    set_keyspace(session, currency, space="raw")
     transactions = session.execute(transaction_search_query[currency],
                                    [expression])
     transactions._fetch_all()
@@ -121,6 +122,30 @@ def query_address_search(currency, expression):
     return addresses
 
 
+def query_label_search(expression_norm_prefix):
+    set_keyspace(session, "", space="tagpacks")
+    labels = session.execute(label_search_query, [expression_norm_prefix])
+    labels._fetch_all()
+    return labels
+
+
+def query_tags(label_norm_prefix, label_norm):
+    set_keyspace(session, "", space="tagpacks")
+    labels = session.execute(tags_query, [label_norm_prefix, label_norm])
+    labels._fetch_all()
+    def makeTagWithCurrency(row):
+        d = gm.Tag(row).__dict__
+        d["currency"] = row.currency
+        return d
+
+    tags = [makeTagWithCurrency(row) for row in labels]
+    return tags
+
+def query_label(label_norm_prefix, label_norm):
+    set_keyspace(session, "", space="tagpacks")
+    label = session.execute(label_query, [label_norm_prefix, label_norm])
+    return gm.Label(label[0]).__dict__ if label else None
+
 def query_address(currency, address):
     set_keyspace(session, currency)
     rows = session.execute(address_query[currency], [address, address[0:5]])
@@ -129,15 +154,20 @@ def query_address(currency, address):
 
 def query_address_cluster(currency, address):
     set_keyspace(session, currency)
-    clusterids = session.execute(address_cluster_query[currency],
-                                 [address, address[0:5]])
+    clusterid = query_address_cluster_id(currency, address)
     ret = {}
-    if clusterids:
-        clusterid = clusterids[0].cluster
+    if clusterid:
         cluster_obj = query_cluster(currency, clusterid)
         ret = cluster_obj.__dict__
     return ret
 
+def query_address_cluster_id(currency, address):
+    set_keyspace(session, currency)
+    clusterids = session.execute(address_cluster_query[currency],
+                                 [address, address[0:5]])
+    if clusterids:
+        return clusterids[0].cluster
+    return None
 
 def query_address_transactions(currency, page_state, address, pagesize, limit):
     set_keyspace(session, currency)
@@ -149,13 +179,14 @@ def query_address_transactions(currency, page_state, address, pagesize, limit):
         query = address_transactions_query
         params = [address, address[0:5], limit]
 
-    if pagesize is not None:
+    if pagesize:
         query[currency].fetch_size = pagesize
-    if page_state is not None:
-        page_state = bytes.fromhex(page_state)
+
+    if page_state:
         rows = session.execute(query[currency], params, paging_state=page_state)
     else:
         rows = session.execute(query[currency], params)
+
     page_state = rows.paging_state
     return page_state, [row for row in rows.current_rows]
 
@@ -165,6 +196,10 @@ def query_address_tags(currency, address):
     tags = session.execute(address_tags_query[currency], [address])
     return [gm.Tag(row).__dict__ for row in tags]
 
+def query_address_with_tags(currency, address):
+    result = query_address(currency, address)
+    result.tags = query_address_tags(currency, address)
+    return result
 
 def query_implicit_tags(currency, address):
     set_keyspace(session, currency)
@@ -185,13 +220,15 @@ def query_address_incoming_relations(currency, page_state, address, pagesize, li
     else:
         query = address_incoming_relations_query
         params = [address[0:5], address, limit]
-    if pagesize is not None:
+
+    if pagesize:
         query[currency].fetch_size = pagesize
-    if page_state is not None:
-        page_state = bytes.fromhex(page_state)
+
+    if page_state:
         rows = session.execute(query[currency], params, paging_state=page_state)
     else:
         rows = session.execute(query[currency], params)
+
     page_state = rows.paging_state
     exchange_rate = gm.ExchangeRate(all_exchange_rates[currency][last_height[currency]])
     relations = [gm.AddressIncomingRelations(row, exchange_rate)
@@ -207,13 +244,15 @@ def query_address_outgoing_relations(currency, page_state, address, pagesize, li
     else:
         query = address_outgoing_relations_query
         params = [address[0:5], address, limit]
+
     if pagesize is not None:
         query[currency].fetch_size = pagesize
+
     if page_state is not None:
-        page_state = bytes.fromhex(page_state)
         rows = session.execute(query[currency], params, paging_state=page_state)
     else:
         rows = session.execute(query[currency], params)
+
     page_state = rows.paging_state
     exchange_rate = gm.ExchangeRate(all_exchange_rates[currency][last_height[currency]])
     relations = [gm.AddressOutgoingRelations(row, exchange_rate)
@@ -246,11 +285,12 @@ def query_cluster_addresses(currency, cluster, page, pagesize, limit):
 
     if pagesize is not None:
         query[currency].fetch_size = pagesize
-    if page is not None:
-        page = bytes.fromhex(page)
+
+    if page:
         rows = session.execute(query[currency], params, paging_state=page)
     else:
         rows = session.execute(query[currency], params)
+
     clusteraddresses = [gm.ClusterAddresses(row, gm.ExchangeRate(all_exchange_rates[currency][last_height[currency]])).__dict__
                         for row in rows.current_rows]
     page = rows.paging_state
@@ -265,13 +305,15 @@ def query_cluster_incoming_relations(currency, page_state, cluster, pagesize, li
     else:
         query = cluster_incoming_relations_query
         params = [cluster, limit]
-    if pagesize is not None:
+
+    if pagesize:
         query[currency].fetch_size = pagesize
-    if page_state is not None:
-        page_state = bytes.fromhex(page_state)
+
+    if page_state:
         rows = session.execute(query[currency], params, paging_state=page_state)
     else:
         rows = session.execute(query[currency], params)
+
     page_state = rows.paging_state
     exchange_rate = gm.ExchangeRate(all_exchange_rates[currency][last_height[currency]])
     relations = [gm.ClusterIncomingRelations(row, exchange_rate) for row in rows.current_rows]
@@ -286,10 +328,11 @@ def query_cluster_outgoing_relations(currency, page_state, cluster, pagesize, li
     else:
         query = cluster_outgoing_relations_query
         params = [cluster, limit]
-    if pagesize is not None:
+
+    if pagesize:
         query[currency].fetch_size = pagesize
-    if page_state is not None:
-        page_state = bytes.fromhex(page_state)
+
+    if page_state:
         rows = session.execute(query[currency], params, paging_state=page_state)
     else:
         rows = session.execute(query[currency], params)
@@ -298,20 +341,72 @@ def query_cluster_outgoing_relations(currency, page_state, cluster, pagesize, li
     relations = [gm.ClusterOutgoingRelations(row, exchange_rate) for row in rows.current_rows]
     return page_state, relations
 
+def query_cluster_search_neighbors(currency, cluster, isOutgoing, category, ids, breadth, depth):
+    set_keyspace(session, currency)
+    if depth <= 0:
+        return []
 
-def set_keyspace(session, currency, raw=False):
-    if currency in currency_mapping:
-        if raw:
-            session.set_keyspace(currency_mapping[currency][0])
+
+    if isOutgoing:
+        (_, rows) = query_cluster_outgoing_relations(currency, None, cluster, breadth, breadth)
+    else:
+        (_, rows) = query_cluster_incoming_relations(currency, None, cluster, breadth, breadth)
+
+    paths = []
+
+    for row in rows:
+        subcluster = row.dstCluster if isOutgoing else row.srcCluster
+        if not subcluster.isdigit():
+            continue
+        match = True
+        tags = query_cluster_tags(currency, subcluster)
+
+        if category != None:
+            # find first occurence of category in tags
+            match = next((True for t in tags if t["category"] == category), False)
+
+        matchingAddresses = []
+        if match and ids != None:
+            matchingAddresses = [id["address"] for id in ids if str(id["cluster"]) == str(subcluster)]
+            match = len(matchingAddresses) > 0
+
+        if match:
+            subpaths = True
         else:
-            session.set_keyspace(currency_mapping[currency][1])
+            subpaths = query_cluster_search_neighbors(currency, subcluster, isOutgoing, category, ids, breadth, depth - 1)
+
+        if not subpaths:
+            continue
+        props = query_cluster(currency, subcluster).__dict__
+        props["tags"] = tags
+        obj = {"node": props, "relation": row.toJson(), "matchingAddresses": []}
+        if subpaths == True:
+            addresses_with_tags = [ query_address_with_tags(currency, address) for address in matchingAddresses ]
+            obj["matchingAddresses"] = [ address for address in addresses_with_tags if address is not None ]
+            subpaths = None
+        obj["paths"] = subpaths
+        paths.append(obj)
+    return paths
+
+def set_keyspace(session, currency=None, space="transformed"):
+    if space == "tagpacks":
+        session.set_keyspace(keyspace_mapping["tagpacks"])
+        return
+
+    if currency in keyspace_mapping:
+        if space == "raw":
+            session.set_keyspace(keyspace_mapping[currency][0])
+        elif space == "transformed":
+            session.set_keyspace(keyspace_mapping[currency][1])
+        else:
+            abort(404, "Keyspace %s not allowed" % space)
     else:
         abort(404, "Currency %s does not exist" % currency)
 
 
 def query_all_exchange_rates(currency, h_max):
     try:
-        set_keyspace(session, currency, raw=True)
+        set_keyspace(session, currency, space="raw")
         session.row_factory = dict_factory
         session.default_fetch_size = None
         print("Loading exchange rates for %s ..." % currency)
@@ -329,7 +424,7 @@ def query_all_exchange_rates(currency, h_max):
 
 
 def query_last_block_height(currency):
-    set_keyspace(session, currency, raw=True)
+    set_keyspace(session, currency, space="raw")
     block_max = 0
     block_inc = 100000
     while True:
@@ -363,10 +458,10 @@ def connect(app):
            blocks_query, cluster_addresses_query, \
            cluster_incoming_relations_query, \
            cluster_outgoing_relations_query, \
-           cluster_query, cluster_tags_query, currency_mapping, \
+           cluster_query, cluster_tags_query, keyspace_mapping, \
            exchange_rate_for_height_query, exchange_rates_query, \
            last_height, session, statistics_query, transaction_search_query, \
-           tx_query, txs_query
+           tx_query, txs_query, label_search_query, label_query, tags_query
 
     cluster = cassandra.cluster.Cluster(app.config["CASSANDRA_NODES"])
     app.logger.debug("Created new Cassandra cluster.")
@@ -374,46 +469,55 @@ def connect(app):
     # set the first keyspace in mapping to the default in order to be able to
     # create the prepared statements; alternative strategy is to not use
     # prepared statements and specify the keyspace in the query string
-    currency_mapping = app.config["MAPPING"]
-    currency = list(currency_mapping.keys())[0]  # just to get the session
-    session = cluster.connect(currency_mapping[currency][1])
+    keyspace_mapping = app.config["MAPPING"]
+    if "tagpacks" in keyspace_mapping.keys() and keyspace_mapping["tagpacks"] == "tagpacks":
+        keyspace_name = "tagpacks"  # it must be "tagpacks"
+    else:
+        abort(404, "Tagpacks keyspace missing")
+
+    session = cluster.connect(keyspace_mapping[keyspace_name])
     session.default_fetch_size = 10
     app.logger.debug("Created new Cassandra session.")
-    for currency in currency_mapping.keys():
-        set_keyspace(session, currency)
-        address_query[currency] = session.prepare("SELECT * FROM address WHERE address = ? AND address_prefix = ?")
-        address_search_query[currency] = session.prepare("SELECT address FROM address WHERE address_prefix = ?")
-        address_transactions_query[currency] = session.prepare("SELECT * FROM address_transactions WHERE address = ? AND address_prefix = ? LIMIT ?")
-        address_transactions_without_limit_query[currency] = session.prepare("SELECT * FROM address_transactions WHERE address = ? AND address_prefix = ?")
-        address_tags_query[currency] = session.prepare("SELECT * FROM address_tags WHERE address = ?")
-        address_cluster_query[currency] = session.prepare("SELECT cluster FROM address_cluster WHERE address = ? AND address_prefix = ?")
-        address_incoming_relations_query[currency] = session.prepare("SELECT * FROM address_incoming_relations WHERE dst_address_prefix = ? AND dst_address = ? LIMIT ?")
-        address_incoming_relations_without_limit_query[currency] = session.prepare("SELECT * FROM address_incoming_relations WHERE dst_address_prefix = ? AND dst_address = ?")
-        address_outgoing_relations_query[currency] = session.prepare("SELECT * FROM address_outgoing_relations WHERE src_address_prefix = ? AND src_address = ? LIMIT ?")
-        address_outgoing_relations_without_limit_query[currency] = session.prepare("SELECT * FROM address_outgoing_relations WHERE src_address_prefix = ? AND src_address = ?")
-        cluster_incoming_relations_query[currency] = session.prepare("SELECT * FROM cluster_incoming_relations WHERE dst_cluster = ? LIMIT ?")
-        cluster_incoming_relations_without_limit_query[currency] = session.prepare("SELECT * FROM cluster_incoming_relations WHERE dst_cluster = ?")
-        cluster_outgoing_relations_query[currency] = session.prepare("SELECT * FROM cluster_outgoing_relations WHERE src_cluster = ? LIMIT ?")
-        cluster_outgoing_relations_without_limit_query[currency] = session.prepare("SELECT * FROM cluster_outgoing_relations WHERE src_cluster = ?")
-        cluster_tags_query[currency] = session.prepare("SELECT * FROM cluster_tags WHERE cluster = ?")
-        cluster_query[currency] = session.prepare("SELECT * FROM cluster WHERE cluster = ?")
-        cluster_addresses_query[currency] = session.prepare("SELECT * FROM cluster_addresses WHERE cluster = ? LIMIT ?")
-        cluster_addresses_without_limit_query[currency] = session.prepare("SELECT * FROM cluster_addresses WHERE cluster = ?")
-        statistics_query[currency] = session.prepare("SELECT * FROM summary_statistics LIMIT 1")
+    label_search_query = session.prepare("SELECT label,label_norm FROM tag_by_label WHERE label_norm_prefix = ? GROUP BY label_norm_prefix, label_norm")
+    label_query = session.prepare("SELECT label_norm, label_norm_prefix, label, COUNT(address) as address_count FROM tag_by_label WHERE label_norm_prefix = ? and label_norm = ? GROUP BY label_norm_prefix, label_norm")
+    tags_query = session.prepare("SELECT * FROM tag_by_label WHERE label_norm_prefix = ? and label_norm = ?")
+    for keyspace_name in keyspace_mapping.keys():
+        if keyspace_name == "tagpacks":
+            continue
+        set_keyspace(session, keyspace_name)
+        address_query[keyspace_name] = session.prepare("SELECT * FROM address WHERE address = ? AND address_prefix = ?")
+        address_search_query[keyspace_name] = session.prepare("SELECT address FROM address WHERE address_prefix = ?")
+        address_transactions_query[keyspace_name] = session.prepare("SELECT * FROM address_transactions WHERE address = ? AND address_prefix = ? LIMIT ?")
+        address_transactions_without_limit_query[keyspace_name] = session.prepare("SELECT * FROM address_transactions WHERE address = ? AND address_prefix = ?")
+        address_tags_query[keyspace_name] = session.prepare("SELECT * FROM address_tags WHERE address = ?")
+        address_cluster_query[keyspace_name] = session.prepare("SELECT cluster FROM address_cluster WHERE address = ? AND address_prefix = ?")
+        address_incoming_relations_query[keyspace_name] = session.prepare("SELECT * FROM address_incoming_relations WHERE dst_address_prefix = ? AND dst_address = ? LIMIT ?")
+        address_incoming_relations_without_limit_query[keyspace_name] = session.prepare("SELECT * FROM address_incoming_relations WHERE dst_address_prefix = ? AND dst_address = ?")
+        address_outgoing_relations_query[keyspace_name] = session.prepare("SELECT * FROM address_outgoing_relations WHERE src_address_prefix = ? AND src_address = ? LIMIT ?")
+        address_outgoing_relations_without_limit_query[keyspace_name] = session.prepare("SELECT * FROM address_outgoing_relations WHERE src_address_prefix = ? AND src_address = ?")
+        cluster_incoming_relations_query[keyspace_name] = session.prepare("SELECT * FROM cluster_incoming_relations WHERE dst_cluster = ? LIMIT ?")
+        cluster_incoming_relations_without_limit_query[keyspace_name] = session.prepare("SELECT * FROM cluster_incoming_relations WHERE dst_cluster = ?")
+        cluster_outgoing_relations_query[keyspace_name] = session.prepare("SELECT * FROM cluster_outgoing_relations WHERE src_cluster = ? LIMIT ?")
+        cluster_outgoing_relations_without_limit_query[keyspace_name] = session.prepare("SELECT * FROM cluster_outgoing_relations WHERE src_cluster = ?")
+        cluster_tags_query[keyspace_name] = session.prepare("SELECT * FROM cluster_tags WHERE cluster = ?")
+        cluster_query[keyspace_name] = session.prepare("SELECT * FROM cluster WHERE cluster = ?")
+        cluster_addresses_query[keyspace_name] = session.prepare("SELECT * FROM cluster_addresses WHERE cluster = ? LIMIT ?")
+        cluster_addresses_without_limit_query[keyspace_name] = session.prepare("SELECT * FROM cluster_addresses WHERE cluster = ?")
+        statistics_query[keyspace_name] = session.prepare("SELECT * FROM summary_statistics LIMIT 1")
 
-        set_keyspace(session, currency, raw=True)
-        tx_query[currency] = session.prepare("SELECT * FROM transaction WHERE tx_prefix = ? AND tx_hash = ?")
-        txs_query[currency] = session.prepare("SELECT * FROM transaction LIMIT ?")
-        transaction_search_query[currency] = session.prepare("SELECT tx_hash from transaction where tx_prefix = ?")
-        block_transactions_query[currency] = session.prepare("SELECT * FROM block_transactions WHERE height = ?")
-        block_query[currency] = session.prepare("SELECT * FROM block WHERE height = ?")
-        blocks_query[currency] = session.prepare("SELECT * FROM block LIMIT ?")
-        exchange_rates_query[currency] = session.prepare("SELECT * FROM exchange_rates LIMIT ?")
-        exchange_rate_for_height_query[currency] = session.prepare("SELECT * FROM exchange_rates WHERE height = ?")
-        block_height_query[currency] = session.prepare("SELECT height FROM exchange_rates WHERE height = ?")
+        set_keyspace(session, keyspace_name, space="raw")
+        tx_query[keyspace_name] = session.prepare("SELECT * FROM transaction WHERE tx_prefix = ? AND tx_hash = ?")
+        txs_query[keyspace_name] = session.prepare("SELECT * FROM transaction LIMIT ?")
+        transaction_search_query[keyspace_name] = session.prepare("SELECT tx_hash from transaction where tx_prefix = ?")
+        block_transactions_query[keyspace_name] = session.prepare("SELECT * FROM block_transactions WHERE height = ?")
+        block_query[keyspace_name] = session.prepare("SELECT * FROM block WHERE height = ?")
+        blocks_query[keyspace_name] = session.prepare("SELECT * FROM block LIMIT ?")
+        exchange_rates_query[keyspace_name] = session.prepare("SELECT * FROM exchange_rates LIMIT ?")
+        exchange_rate_for_height_query[keyspace_name] = session.prepare("SELECT * FROM exchange_rates WHERE height = ?")
+        block_height_query[keyspace_name] = session.prepare("SELECT height FROM exchange_rates WHERE height = ?")
 
-        last_height[currency] = query_last_block_height(currency)
-        all_exchange_rates[currency] = query_all_exchange_rates(currency,
-                                                                last_height[currency])
+        last_height[keyspace_name] = query_last_block_height(keyspace_name)
+        all_exchange_rates[keyspace_name] = query_all_exchange_rates(keyspace_name,
+                                                                last_height[keyspace_name])
 
     app.logger.debug("Created prepared statements")