+++ app/wsgi.py
@@ -1,6 +1,6 @@
+from uwsgidecorators import postfork
 from graphsenserest import app as application
 from graphsensedao import connect
-from uwsgidecorators import postfork
 
 
 @postfork
@@ -11,5 +11,5 @@ def postfork_connect():
 if __name__ == "__main__":
     # deactivate debug and multiple processes in production
     # because of memory usage and security
-    # host='0.0.0.0',port='8888'
+    # host="0.0.0.0",port="8888"
     application.run(debug=False, processes=1)