+++ docker/build.sh
@@ -1,1 +1,1 @@
-docker build -t graphsense-rest .
+docker build --build-arg rest_user=admin --build-arg rest_passwd=graphsense -t graphsense-rest .